Github_Integration_Backend/requirements.txt
asgiref==3.5.1
Django==4.0.4
django-cors-headers==3.11.0
djangorestframework==3.13.1
pytz==2022.1
sqlparse==0.4.2